Adaptation to the new version of learncpp.com website
With this new version of the website what I covered is basically everything till chapter 21.x (I miss the project 21.y).
I read the lessons marked as `new` (16.[1,3,4,6,11], 17.[2,3,5,6,13]). The impression I have is that the lessons marked as `updated` got small modifications and some rearrangement of the topics, but not essential modifications, thus I decided to go forward from starting from lecture 21.y